code/test-environment/requirements.txt
kafka-python==2.0.2
python-dotenv
matplotlib
pandas
tqdm